refactor: merge ChimeMeetingSession
